launchpad: Reorganise fee files
