+++ packages/launchpad/src/tx.spec.ts
@@ -3,8 +3,8 @@ import { StdSignature } from "@cosmjs/amino";
 
 import { coins } from "./coins";
 import { makeSignDoc } from "./encoding";
+import { StdFee } from "./fee";
 import { makeStdTx } from "./tx";
-import { StdFee } from "./types";
 
 describe("tx", () => {
   describe("makeStdTx", () => {
