--- packages/launchpad/src/types.ts
@@ -1,7 +1,0 @@
-/* eslint-disable @typescript-eslint/naming-convention */
-import { Coin } from "./coins";
-
-export interface StdFee {
-  readonly amount: readonly Coin[];
-  readonly gas: string;
-}
